Refactor: TasteGraphからform用の内容を分離した
今後、indexでのグラフ画像表示をするために、共通部を抜き出した。
