+++ app/views/sakes/_form.html.erb
@@ -232,3 +232,4 @@
 <% end %>
 
 <%= javascript_pack_tag "taste_graph" %>
+<%= javascript_pack_tag "form_graph" %>
